kite-lxc: improvements around base methods
